Individual experiments
Now experiments are read in individually and save as separate files with the date and the well they are assigned to. These are looped through, calling run_experiments.py for each one. As the experiments run, the well_status.json and the experiment.json files are updated.

When reading in experiments, the desired well is checked against well_status.json and if that well is not available the next available well is assigned to the experiment. If there are no available wells the instructions are not moved to archive and a print out tells the user that not all experiments have been assigned.

Additionally changes were made to add functions to allow all of this as well as folders for organization