Use the config file's DB path
For singlurlar versions of plural commands, simply pass the singular argument as a list to the plural to reduce code duplication. The operations should be the same.

An exception to this exists with the update experiment function, these are kept seperate where the plural form only accepts ExperimentBase objects while the singular can take an int as an experiment ID